OPENMEETINGS-4 Microphone rooms I made some changes to the code:
You've placed the bottom bar in the video view on y="${this.height}" so that it does not cover the webcam video, however I would prefer having the views inside the width and height of the view instead of attaching it to the bottom. It seems to me like otherwise the drop-shadow of the video-pods doesn't work as it is calced/based on the width/height of the video-pod's view.

So I changed:
videoObjectBroadcast.lzx
videoObjectPlayBroadcast.lzx
method: setDefaultVideoSize(w,h)
to calc additionally the bottom bar size into the resulting video size

setDefaultVideoSize(w,h) is triggered by editRecordStream.lzx Line 135

By using that we don't need "_resizeviewMicro" view and only use the resizeview 

git-svn-id: https://svn.apache.org/repos/asf/incubator/openmeetings/trunk/singlewebapp@1221690 13f79535-47bb-0310-9956-ffa450edef68